aqua-ai-v2/requirements.txt
streamlit
streamlit-webrtc
transformers
torch
pillow
opencv-python
av